feat: moved prompts to separate files
