+++ kaizen/llms/prompts/general.py
@@ -1,0 +1,1 @@
+BASIC_SYSTEM_PROMPT = "You are an helpful AI Assistant"
