+++ kaizen/reviewer/code_review.py
@@ -1,7 +1,7 @@
-from kaizen.helpers import output, parser
 from typing import Optional, List, Dict
+from kaizen.helpers import output, parser
 from kaizen.llms.provider import LLMProvider
-from kaizen.llms.prompts import (
+from kaizen.llms.prompts.code_review import (
     CODE_REVIEW_PROMPT,
     CODE_REVIEW_SYSTEM_PROMPT,
     PR_DESCRIPTION_PROMPT,
